Move func show menber -> friend
