+++ include/GameBoard.hpp
@@ -2,7 +2,6 @@
 #define GAME_BOARD_H_
 
 #include <stdexcept>
-#include <iostream>
 
 #include "BoardBase.hpp"
 
@@ -12,14 +11,12 @@ class GameBoard : public BoardBase {
 public:
   GameBoard(const Position size = Position{8, 8});
   const GameBoard& operator=(const GameBoard& src) = delete;
-  void show() const;
   void init();
   Stone get_active_stone() const;
   bool can_continue() const;
   void put_dots();
   void remove_dots();
 };
-char to_char(Stone stone);
 };
 
 #endif // GAME_BOARD_H_
