+++ src/GameBoard.cpp
@@ -7,34 +7,6 @@ GameBoard::GameBoard(const Position size)
 {
 }
 
-char to_char(Stone stone)
-{
-  switch (stone) {
-  case Stone::Space: return ' ';
-  case Stone::White: return 'O';
-  case Stone::Black: return 'X';
-  case Stone::Dot:   return '*';
-  default: throw std::invalid_argument{"In to_char:Cannot convert"};
-  }  
-}
-
-void GameBoard::show() const
-{
-  for (int i {0}; i < width()+1; i++) std::cout << "--";
-  std::cout << std::endl;
-  std::cout << "  " << std::flush;
-  for (int i {0}; i < width(); std::cout.put(' '), i++)
-    std::cout << i+1 << std::flush;
-  std::cout << std::endl;
-  for (PlaneVector::Point column {0}; column < height(); std::cout.put('\n'), column++) {
-    std::cout << column+1 << ' ' << std::flush;
-    for (PlaneVector::Point row {0}; row < width(); std::cout.put(' '), row++)
-      std::cout.put(to_char(board_[get_access_num(row, column)]));
-  }
-  for (int i {0}; i < width()+1; i++) std::cout << "--";
-  std::cout << std::endl;
-}
-
 void GameBoard::init()
 {
   board_ = Stone::Space;
